requirements.txt
beautifulsoup4>=4.7.1
erddapy>=0.5.0
numpy>=1.16.4
pandas>=0.25.0
Pydap>=3.2.2
pytz>=2019.1
PyYAML==5.1.2
pyyaml-include==1.1.1.1
requests>=2.22.0
urllib3>=1.25.3
git+git://github.com/cioos-siooc/metadata-xml.git
